# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/subscriptions/me": {
      "get": {
        "operationId": "Subscriptions_findMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired token (`unauthenticated`)."
          },
          "403": {
            "description": "The account does not administer a club (`club_required`)."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Returns the effective subscription for the current club",
        "tags": [
          "Clubs"
        ]
      }
    },
    "/subscriptions/me/upgrade-quote": {
      "get": {
        "operationId": "Subscriptions_quoteUpgrade",
        "parameters": [
          {
            "name": "targetPlan",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pro"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImmediateUpgradeQuoteDto"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed (`validation`)."
          },
          "409": {
            "description": "An active paid Basic period with verified charge history is required."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Quotes an immediate prorated Basic-to-Pro upgrade",
        "tags": [
          "Clubs"
        ]
      }
    },
    "/subscriptions/me/downgrade-quote": {
      "get": {
        "operationId": "Subscriptions_quoteDowngrade",
        "parameters": [
          {
            "name": "targetPlan",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "basic"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDowngradeQuoteResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Quotes Pro-to-Basic for the next paid renewal",
        "tags": [
          "Clubs"
        ]
      }
    },
    "/subscriptions/me/downgrade": {
      "post": {
        "operationId": "Subscriptions_scheduleDowngrade",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanDowngradeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDowngradeQuoteResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "The subscription cannot be downgraded now."
          },
          "503": {
            "description": "Mercado Pago could not confirm the recurring amount update."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Schedules Pro-to-Basic after the paid Pro period",
        "tags": [
          "Clubs"
        ]
      }
    },
    "/subscriptions/me/upgrade": {
      "post": {
        "operationId": "Subscriptions_scheduleUpgrade",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUpgradeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImmediateUpgradeCheckoutDto"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed (`validation`)."
          },
          "409": {
            "description": "The subscription is not eligible, another plan change is in progress, or the quoted amount changed (`upgrade_quote_changed`)."
          },
          "503": {
            "description": "The preference creation or recovery is unavailable."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Starts a one-time prorated Basic-to-Pro payment",
        "tags": [
          "Clubs"
        ]
      }
    },
    "/subscriptions/me/checkouts": {
      "post": {
        "operationId": "Subscriptions_createCheckout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed (`validation`)."
          },
          "401": {
            "description": "Missing, invalid, or expired token (`unauthenticated`)."
          },
          "403": {
            "description": "The account does not administer a club (`club_required`)."
          },
          "409": {
            "description": "An active Basic subscription must use `/subscriptions/me/upgrade` (`subscription_upgrade_required`); another active subscription cannot start a second checkout until its paid period ends. A checkout may also be pending or in progress."
          },
          "503": {
            "description": "Billing configuration or checkout recovery is unavailable (`billing_not_configured`, `billing_checkout_recovery_required`)."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Starts or reuses the monthly Mercado Pago recurring checkout",
        "tags": [
          "Clubs"
        ]
      }
    },
    "/subscriptions/me/resume": {
      "post": {
        "operationId": "Subscriptions_resumeMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Subscription renewal was resumed."
          },
          "409": {
            "description": "The paused subscription period has ended (`paused_subscription_period_ended`)."
          },
          "503": {
            "description": "Mercado Pago is unavailable (`billing_provider_unavailable`)."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Resumes a paused Mercado Pago recurring subscription",
        "tags": [
          "Clubs"
        ]
      }
    },
    "/subscriptions/me/cancel": {
      "post": {
        "operationId": "Subscriptions_cancelMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Subscription renewal was paused."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Stops renewal while preserving the paid subscription period",
        "tags": [
          "Clubs"
        ]
      }
    },
...
  },
  "info": {
    "title": "dupla API",
    "description": "API de dupla — SaaS de torneos de pádel para clubes.\n\nLos endpoints se agrupan por dominio de negocio. Su clase de acceso es un\ncontrato separado: `club` (JWT de staff y scoping por el club del usuario),\n`public` (solo lectura y sin auth), `platform` (entidades globales sin\n`club_id`) u `ops` (infraestructura).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token emitido por la API (Passport + JWT)."
      }
    },
    "schemas": {
...
      "PendingUpgradeDto": {
        "type": "object",
        "properties": {
          "state": {
            "enum": [
              "creating",
              "pending",
              "paid",
              "review_required"
            ],
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "amount": {
            "type": "string",
            "example": "7500.00"
          },
          "currency": {
            "type": "string",
            "example": "ARS"
          },
          "checkoutUrl": {
            "type": "string",
            "nullable": true
          },
          "periodEndsAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "state",
          "reference",
          "amount",
          "currency",
          "checkoutUrl",
          "periodEndsAt"
        ]
      },
      "PlanDowngradeQuoteResponseDto": {
        "type": "object",
        "properties": {
          "targetPlan": {
            "type": "string",
            "enum": [
              "basic"
            ]
          },
          "amount": {
            "type": "string",
            "example": "10000.00"
          },
          "currency": {
            "type": "string",
            "example": "ARS"
          },
          "effectiveAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-10-21T00:00:00.000Z"
          }
        },
        "required": [
          "targetPlan",
          "amount",
          "currency",
          "effectiveAt"
        ]
      },
      "SubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "plan": {
            "enum": [
              "free",
              "basic",
              "pro"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "active",
              "past_due",
              "canceled"
            ],
            "type": "string"
          },
          "maxTournaments": {
            "type": "number",
            "example": 1
          },
          "currentPeriodEndsAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "example": "2026-10-21T00:00:00.000Z"
          },
          "renewsAutomatically": {
            "type": "boolean",
            "example": true
          },
          "pendingUpgrade": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PendingUpgradeDto"
              }
            ]
          },
          "pendingDowngrade": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlanDowngradeQuoteResponseDto"
              }
            ]
          }
        },
        "required": [
          "plan",
          "status",
          "maxTournaments",
          "currentPeriodEndsAt",
          "renewsAutomatically",
          "pendingUpgrade",
          "pendingDowngrade"
        ]
      },
      "ImmediateUpgradeQuoteDto": {
        "type": "object",
        "properties": {
          "targetPlan": {
            "type": "string",
            "enum": [
              "pro"
            ]
          },
          "amount": {
            "type": "string",
            "example": "7500.00"
          },
          "recurringAmount": {
            "type": "string",
            "example": "25000.00"
          },
          "currency": {
            "type": "string",
            "example": "ARS"
          },
          "periodEndsAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-10-21T00:00:00.000Z"
          }
        },
        "required": [
          "targetPlan",
          "amount",
          "recurringAmount",
          "currency",
          "periodEndsAt"
        ]
      },
      "PlanDowngradeRequestDto": {
        "type": "object",
        "properties": {
          "targetPlan": {
            "type": "string",
            "enum": [
              "basic"
            ]
          }
        },
        "required": [
          "targetPlan"
        ]
      },
      "CreateUpgradeRequestDto": {
        "type": "object",
        "properties": {
          "targetPlan": {
            "type": "string",
            "enum": [
              "pro"
            ]
          },
          "expectedAmount": {
            "type": "string",
            "pattern": "^(0|[1-9]\\d*)\\.\\d{2}$",
            "example": "7500.00"
          }
        },
        "required": [
          "targetPlan",
          "expectedAmount"
        ]
      },
      "ImmediateUpgradeCheckoutDto": {
        "type": "object",
        "properties": {
          "targetPlan": {
            "type": "string",
            "enum": [
              "pro"
            ]
          },
          "amount": {
            "type": "string",
            "example": "7500.00"
          },
          "recurringAmount": {
            "type": "string",
            "example": "25000.00"
          },
          "currency": {
            "type": "string",
            "example": "ARS"
          },
          "periodEndsAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-10-21T00:00:00.000Z"
          },
          "reference": {
            "type": "string",
            "example": "9fd80e8b-cce4-475e-a80e-a900a8ff3be8"
          },
          "checkoutUrl": {
            "type": "string",
            "example": "https://www.mercadopago.com.ar/checkout/v1/redirect?pref_id=..."
          },
          "reused": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "targetPlan",
          "amount",
          "recurringAmount",
          "currency",
          "periodEndsAt",
          "reference",
          "checkoutUrl",
          "reused"
        ]
      },
      "CreateCheckoutDto": {
        "type": "object",
        "properties": {
          "plan": {
            "enum": [
              "basic",
              "pro"
            ],
            "type": "string"
          }
        },
        "required": [
          "plan"
        ]
      },
      "CheckoutResponseDto": {
        "type": "object",
        "properties": {
          "plan": {
            "enum": [
              "free",
              "basic",
              "pro"
            ],
            "type": "string"
          },
          "reference": {
            "type": "string",
            "format": "uuid"
          },
          "checkoutUrl": {
            "type": "string",
            "format": "uri"
          },
          "reused": {
            "type": "boolean",
            "description": "true when an existing pending checkout was reused"
          }
        },
        "required": [
          "plan",
          "reference",
          "checkoutUrl",
          "reused"
        ]
      },
...
    }
  }
}